Moved to 'stdint.h' types only.

+++ hash_map.h
@@ -22,6 +22,6 @@ uint64_t hash_map_length(struct hash_map *map);
 uint64_t hash_map_capacity(struct hash_map *map);
 struct hash_map_iter *hash_map_iter_create(struct hash_map *map);
 void hash_map_iter_destroy(struct hash_map_iter *iter);
-int hash_map_iter_next(struct hash_map_iter *iter, char **key, void **val);
+int64_t hash_map_iter_next(struct hash_map_iter *iter, char **key, void **val);
 
 #endif
